Fix bug in time window formulation.
This seems to work, but still not sure.
Problem previously was that the time leaving each location was constrained, but the time returning to location 1 was not.
Can't just constrain everything from 1:N, as then we get a recursive addition.
So solution is to have two time variables for the first location, one for arriving and one for leaving.